feat: implement thing management and user authentication controllers

+++ backend/controllers/stuff.js
@@ -1,6 +1,6 @@
 const Thing = require('../models/things')
 const { cleanImageUrl, cleanThingImageUrl } = require('../utils/cleanImageUrl')
- 
+ const fs = require('fs')
 exports.createThing = (req, res, next) => {
     const thingObject = JSON.parse(req.body.thing);
     delete thingObject._id;
@@ -13,7 +13,6 @@ exports.createThing = (req, res, next) => {
     thing.save()
         .then(() => res.status(201).json({ message: "donnees postees avec succes " }))
         .catch((err) => {
-            console.log("Erreur d'enregistrement MongoDB :", err); // Affiche l'erreur exacte dans le terminal
             res.status(400).json({ error: err.message || err });
         });
 };
@@ -22,12 +21,34 @@ exports.updateThing = (req, res, next) => {
     const thingObject = req.file ? { ...JSON.parse(req.body.thing), 
     imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}` } : { ...req.body }
     delete thingObject.userId;
+    Thing.findOne({_id:req.params.id})
+    .then((thing) => {
+        if(thing.userId != req.auth.userId){
+            res.status(401).json({message:"acces non autorise"})
+        }else{
+            Thing.updateOne({_id:req.params.id}, {...thingObject, _id:req.params.id})
+            .then(() => res.status(200).json({message:'objet modifie avec succes'}))
+            .catch((error) => res.status(400).json(error))
+        }
+    })
+    .catch(err => res.status(400).json(err))
 
-}
+} 
 
 exports.deleteThing = (req, res, next) => {
-    Thing.deleteOne({_id:req.params.id})
-    .then(() => res.status(200).json({message:"article supprime avec succes"}))
+    Thing.findOne({_id:req.params.id})
+    .then(thing => {
+        if(thing.userId != req.auth.userId){
+            res.status(400).json({message:"erreur de suppression"})
+        }else{
+            const filename = thing.imageUrl.split('/images/')[1]
+            fs.unlink(`images/${filename}`, () => {
+                Thing.deleteOne({_id:req.params.id})
+                .then(() => res.status(200).json({message: 'objet supprimé avec succes'}))
+                .catch(err => res.status(400).json({err}))
+            })
+        }
+    })
     .catch((err) => res.status(400).json({error: err}))
 }
